refactor(STADTPULS-355): move useUserTokens utilities to requests folder
